# src/apiHandlers/swaggerHandlers/swagger.yaml
definitions:
  APIErrorCode:
    enum:
    - 2100
    - 2101
    - 2102
    - 2103
    - 2104
    - 2105
    - 2106
    - 2107
    - 2108
    - 2109
    - 2110
    - 2111
    - 2112
    - 2113
    - 2114
    - 2115
    - 2116
    - 2117
    - 2118
    - 2119
    - 2120
    - 2121
    - 2122
    - 2123
    - 2124
    - 2125
    - 2126
    - 2127
    - 2128
    - 2129
    - 2130
    - 2131
    - 2132
    - 2133
    - 2134
    - 2135
    - 2136
    - 2137
    - 2138
    - 2139
    - 2140
    - 2141
    - 2142
    type: integer
    x-enum-varnames:
    - ErrCodeMalformedJWT
    - ErrCodeInvalidJWT
    - ErrCodeInvalidBodyData
    - ErrCodeInvalidUsernamePass
    - ErrCodeInvalidAuth
    - ErrCodePermissionDenied
    - ErrCodeInvalidInputPass
    - ErrCodeUsernameExists
    - ErrCodeInternalServerError
    - ErrCodeInvalidFileData
    - ErrCodeInvalidPhoneNumber
    - ErrCodePhoneNumberAlreadyImported
    - ErrCodeInvalidUsername
    - ErrCodeNoPhonesDonated
    - ErrCodeAgentNotConnected
    - ErrCodeInvalidPagination
    - ErrCodeMaxContactImportLimit
    - ErrCodePhoneNumberNotFound
    - ErrCodeParameterRequired
    - ErrCodeUserBanned
    - ErrCodeLabelInfoNotFound
    - ErrCodeLabelAlreadyApplied
    - ErrCodeLabelAlreadyExistsByName
    - ErrCodeTooManyChatLabelInfo
    - ErrCodeLabelNameTooLong
    - ErrCodeLabelDescriptionTooLong
    - ErrCodeInvalidColor
    - ErrCodeLabelNotApplied
    - ErrCodeCannotDeleteBuiltInLabel
    - ErrCodeDuplicatePhoneNumber
    - ErrCodePhoneNotWorking
    - ErrCodeInvalidPmsPass
    - ErrCodeInvalidAgentId
    - ErrCodeInvalidAppSettingName
    - ErrCodeAppSettingNotFound
    - ErrCodeTextEmpty
    - ErrCodeTextTooLong
    - ErrCodeInvalidClientRId
    - ErrCodeInvalidCaptcha
    - ErrCodeQueryParameterNotProvided
    - ErrCodeTooManyPasswordChangeAttempts
    - ErrCodeRequestExpired
    - ErrCodeInvalidEmail
  AuthResult:
    properties:
      access_token:
        type: string
      expiration:
        type: integer
      full_name:
        type: string
      refresh_token:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  BanUserData:
    properties:
      ban_reason:
        type: string
      is_banned:
        type: boolean
      user_id:
        type: string
    type: object
  BanUserResult:
    properties:
      ban_reason:
        type: string
      is_banned:
        type: boolean
      user_id:
        type: string
    type: object
  ChangePasswordData:
    properties:
      lang:
        default: en
        type: string
      new_password:
        type: string
      user_id:
        type: string
    type: object
  ConfirmChangePasswordData:
    properties:
      new_password:
        type: string
      rq_id:
        type: string
      rt_hash:
        type: string
      rt_param:
        type: string
      rt_verifier:
        type: string
    type: object
  CreateNewTopicData:
    properties:
      topicName:
        type: string
    type: object
  CreateNewTopicResult:
    properties:
      topicID:
        type: integer
      topicName:
        type: string
    type: object
  CreateUserData:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role:
        $ref: '#/definitions/UserRole'
      setup_completed:
        type: boolean
      user_address:
        type: string
      user_id:
        type: string
    type: object
  CreateUserResult:
    properties:
      email:
        type: string
      full_name:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  EditUserData:
    properties:
      email:
        type: string
      full_name:
        type: string
      phone_number:
        type: string
      user_address:
        type: string
      user_id:
        type: string
    type: object
  EditUserResult:
    properties:
      email:
        type: string
      full_name:
        type: string
      role:
        $ref: '#/definitions/UserRole'
      user_id:
        type: string
    type: object
  EndpointError:
    properties:
      code:
        $ref: '#/definitions/APIErrorCode'
      date:
        type: string
      message:
        type: string
      origin:
        type: string
    type: object
  EndpointResponse:
    properties:
      error:
        $ref: '#/definitions/EndpointError'
      result: {}
      success:
        type: boolean
    type: object
  GetMeResult:
    properties:
      full_name:
        type: string
      role:
        $ref: '#/definitions/UserRole'
      user_id:
        type: string
    type: object
  GetUserInfoResult:
    properties:
      ban_reason:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      is_banned:
        type: boolean
      role:
        $ref: '#/definitions/UserRole'
      user_id:
        type: string
    type: object
  LoginData:
    properties:
      captcha_answer:
        type: string
      captcha_id:
        type: string
      client_rid:
        type: string
      password:
        type: string
      user_id:
        type: string
    type: object
  LoginResult:
    properties:
      access_token:
        type: string
      expiration:
        type: integer
      full_name:
        type: string
      refresh_token:
        type: string
      role:
        $ref: '#/definitions/UserRole'
      user_id:
        type: string
    type: object
  SearchUserData:
    properties:
      limit:
        minimum: 1
        type: integer
      offset:
        type: integer
      query:
        type: string
    type: object
  SearchUserResult:
    properties:
      users:
        items:
          $ref: '#/definitions/SearchedUserInfo'
        type: array
    type: object
  SearchedUserInfo:
    properties:
      ban_reason:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      is_banned:
        type: boolean
      role:
        $ref: '#/definitions/UserRole'
      user_id:
        type: string
    type: object
  UserRole:
    description: UserRole is the role of the user.
    enum:
    - owner
    - admin
    - student
    - teacher
    - ""
    type: string
    x-enum-varnames:
    - UserRoleOwner
    - UserRoleAdmin
    - UserRoleStudent
    - UserRoleTeacher
    - UserRoleUnknown
  captchaHandlers.GetCaptchaResult:
    properties:
      captcha:
        type: string
      captcha_id:
        type: string
      client_r_id:
        type: string
    type: object
info:
  contact: {}
  description: This is the API for the ExamSphere system
  title: ExamSphere API
  version: "1.0"
paths:
  /api/v1/captcha/generate:
    get:
      description: Allows a client (with Client-R-ID) to generate a captcha
      operationId: GenerateCaptchaV1
      parameters:
      - description: Client-R-ID
        in: header
        name: Client-R-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  $ref: '#/definitions/captchaHandlers.GetCaptchaResult'
              type: object
      summary: Get a captcha
      tags:
      - User
  /api/v1/user/ban:
    post:
      consumes:
      - application/json
      description: Allows a user to ban another user
      operationId: banUserV1
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ban user data
        in: body
        name: banUserData
        required: true
        schema:
          $ref: '#/definitions/BanUserData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  $ref: '#/definitions/BanUserResult'
              type: object
      summary: Ban a user
      tags:
      - User
  /api/v1/user/changePassword:
    post:
      consumes:
      - application/json
      description: Allows a user to change a user's password.
      operationId: changePasswordV1
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Change password data
        in: body
        name: changePasswordData
        required: true
        schema:
          $ref: '#/definitions/ChangePasswordData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  type: boolean
              type: object
      summary: Change a user's password
      tags:
      - User
  /api/v1/user/confirmChangePassword:
    post:
      consumes:
      - application/json
      description: Allows a user to confirm changing their own's password (from redirected
        page)
      operationId: confirmChangePasswordV1
      parameters:
      - description: Confirm change password data
        in: body
        name: confirmChangePasswordData
        required: true
        schema:
          $ref: '#/definitions/ConfirmChangePasswordData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  type: boolean
              type: object
      summary: Confirm changing your own's password
      tags:
      - User
  /api/v1/user/create:
    post:
      consumes:
      - application/json
      description: Allows a user to create a new user
      operationId: createUserV1
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Create user data
        in: body
        name: createUserData
        required: true
        schema:
          $ref: '#/definitions/CreateUserData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  $ref: '#/definitions/CreateUserResult'
              type: object
      summary: Create a new user
      tags:
      - User
  /api/v1/user/edit:
    post:
      consumes:
      - application/json
      description: Allows a user to edit another user. Users are not allowed to edit
        their own information.
      operationId: editUserV1
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Edit user data
        in: body
        name: editUserData
        required: true
        schema:
          $ref: '#/definitions/EditUserData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  $ref: '#/definitions/EditUserResult'
              type: object
      summary: Edit a user's basic information
      tags:
      - User
  /api/v1/user/info:
    get:
      description: Allows a user to get another user's information by their user ID
      operationId: getUserInfoV1
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  $ref: '#/definitions/GetUserInfoResult'
              type: object
      summary: Get a user's information
      tags:
      - User
  /api/v1/user/login:
    post:
      consumes:
      - application/json
      description: Allows a user to login to the system and obtain access/refresh
        tokens
      operationId: loginV1
      parameters:
      - description: Login data
        in: body
        name: loginData
        required: true
        schema:
          $ref: '#/definitions/LoginData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  $ref: '#/definitions/LoginResult'
              type: object
      summary: Login to the system
      tags:
      - User
  /api/v1/user/me:
    get:
      description: Allows a user to get their own information
      operationId: getMeV1
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  $ref: '#/definitions/GetMeResult'
              type: object
      summary: Get the user's information
      tags:
      - User
  /api/v1/user/reAuth:
    post:
      description: Allows a user to refresh their access token
      operationId: reAuthV1
      parameters:
      - description: Refresh token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  $ref: '#/definitions/AuthResult'
              type: object
      summary: Refresh the access token
      tags:
      - User
  /api/v1/user/search:
    post:
      consumes:
      - application/json
      description: Allows a user to search for users
      operationId: searchUserV1
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Search user data
        in: body
        name: searchUserData
        required: true
        schema:
          $ref: '#/definitions/SearchUserData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/EndpointResponse'
            - properties:
                result:
                  $ref: '#/definitions/SearchUserResult'
              type: object
      summary: Search users
      tags:
      - User
swagger: "2.0"